==> 01 preprocess.ipynb <==
# Import packages
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler,LabelEncoder
import pickle

## Load the dataset
data = pd.read_csv("dataset/Churn_Modelling.csv")
data.head()

## Preprocess the data
### Drop irrelevant columns
data = data.drop(['RowNumber','CustomerId','Surname'], axis=1)
data

## Encode categorical variables
le_gender = LabelEncoder()
data['Gender'] = le_gender.fit_transform(data['Gender'])
data

## Onehot encoding
from sklearn.preprocessing import OneHotEncoder
ohe_geography = OneHotEncoder()
geo_encoder = ohe_geography.fit_transform(data[['Geography']]).toarray()
geo_encoder

ohe_geography.get_feature_names_out(['Geography'])

geo_encoded_df = pd.DataFrame(geo_encoder, columns=ohe_geography.get_feature_names_out(['Geography']))
geo_encoded_df

## Combine one hot encoder columns with the original data
data = pd.concat([data.drop('Geography', axis=1), geo_encoded_df], axis=1)
data.head()

## Save the encoders and scaler
with open('models/label_encoder_gender.pkl','wb') as file:
    pickle.dump(le_gender,file)

with open('models/onehot_encoder_geo.pkl','wb') as file:
    pickle.dump(ohe_geography,file)

data.head()

## DiVide the dataset into indepent and dependent features
X = data.drop('Exited', axis=1)
y = data['Exited']

## Split the data in training and tetsing sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

## Scale these features
scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

X_train

with open('models/scaler.pkl','wb') as file:
    pickle.dump(scaler,file)

data

### ANN Implementation

import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense
from tensorflow.keras.callbacks import EarlyStopping, TensorBoard
import datetime

(X_train.shape[1],)

## Build Our ANN Model

model=Sequential([
    Dense(64,activation='relu',input_shape=(X_train.shape[1],)), ## HL1 Connected wwith input layer
    Dense(32,activation='relu'), ## HL2
    Dense(1,activation='sigmoid')  ## output layer
]
)

model.summary()

import tensorflow
opt=tensorflow.keras.optimizers.Adam(learning_rate=0.01)
loss=tensorflow.keras.losses.BinaryCrossentropy()
loss

## compile the model
model.compile(optimizer=opt, loss="binary_crossentropy", metrics=['accuracy'])

## Set up the Tensorboard
from tensorflow.keras.callbacks import EarlyStopping, TensorBoard

log_dir="logs/fit/" + datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
tensorflow_callback = TensorBoard(log_dir=log_dir,histogram_freq=1)

## Set up Early Stopping
early_stopping_callback = EarlyStopping(monitor='val_loss', patience=10, restore_best_weights=True)

### Train the model
history = model.fit(
    X_train, y_train, validation_data = (X_test,y_test), epochs=100, callbacks=[tensorflow_callback, early_stopping_callback]
)

model.save('models/ANN_model.h5')

## Load Tensorboard Extension
%load_ext tensorboard

%tensorboard --logdir logs/fit
